site_django-main/requirements.txt
<<<<<<< HEAD
absl-py==2.0.0
anyio==4.0.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==3.1.1
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.2
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.15
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-sslserver==0.22
djangorestframework==3.15.2
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.3
fasteners==0.14.1
fastjsonschema==2.18.1
filelock==3.0.12
flatbuffers==23.5.26
fsspec==2023.10.0
future==0.18.2
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.15.2
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-client==8.5.0
jupyter-console==6.6.3
jupyter-core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter-server==2.9.1
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
keras==2.13.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.6
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.5.1
MarkupSafe==2.1.3
mistune==3.0.2
monotonic==1.5
more-itertools==4.2.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
netifaces==0.10.4
networkx==3.1
notebook==6.0.3
notebook-shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pbr==6.1.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
prometheus-client==0.7.1
prompt-toolkit==3.0.39
protobuf==4.25.0
psycopg==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pygame==2.5.2
Pygments==2.16.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.7
python-magic==0.4.16
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
rsa==4.9
SecretStorage==2.3.1
Send2Trash==1.8.2
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.1
ssh-import-id==5.10
sympy==1.12
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.4.4
testresources==2.0.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
traitlets==5.13.0
triton==2.1.0
typing-extensions==4.5.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.6.4
werkzeug==3.0.1
widgetsnbextension==2.0.0
wrapt==1.15.0
xkit==0.0.0
zipp==1.0.0
=======
asgiref==3.8.1
Django==4.2.15
sqlparse==0.5.1
typing-extensions==4.12.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
>>>>>>> parte2/main
